chore: Remove ExpandUser function
It was a naive implementation anyway and isn't something I want to
support as part of this library (there are better implementations out
there).

The one case where we need to do naive expansion has been replaced with
a private function that has a name that clarifies it is naive.